+++ tests/testthat/test-position_likert.R
@@ -21,17 +21,17 @@ test_that("position_likert()", {
 
   p <- ggplot(diamonds) +
     aes(y = clarity, fill = cut) +
-    geom_bar(position = "likert_count") +
+    geom_bar(position = "diverging") +
     scale_x_continuous(label = label_number_abs()) +
     scale_fill_brewer(palette = "PiYG")
 
   vdiffr::expect_doppelganger(
-    "position_likert_count() base",
+    "position_diverging() base",
     p
   )
 
   vdiffr::expect_doppelganger(
-    "position_likert_count() facet",
+    "position_diverging() facet",
     p + facet_grid(~ price > 2500)
   )
 
@@ -46,10 +46,10 @@ test_that("position_likert()", {
 
   p <- ggplot(diamonds) +
     aes(y = clarity, fill = cut) +
-    geom_bar(position = position_likert_count(reverse = TRUE))
+    geom_bar(position = position_diverging(reverse = TRUE))
 
   vdiffr::expect_doppelganger(
-    "position_likert_count() reverse",
+    "position_diverging() reverse",
     p
   )
 
@@ -85,6 +85,39 @@ test_that("position_likert()", {
 
   vdiffr::expect_doppelganger(
     "position_likert() exclude_fill_values",
+    p
+  )
+})
+
+test_that("geom_diverging() & associates", {
+  library(ggplot2)
+  p <-
+    ggplot(diamonds) +
+    aes(y = clarity, fill = cut) +
+    geom_diverging() +
+    geom_diverging_text(aes(color = after_scale(hex_bw(.data$fill))))
+  vdiffr::expect_doppelganger(
+    "geom_diverging and geom_diverging_text",
+    p
+  )
+
+  p <-
+    ggplot(diamonds) +
+    aes(y = clarity, fill = cut) +
+    geom_likert() +
+    geom_likert_text(aes(color = after_scale(hex_bw(.data$fill))))
+  vdiffr::expect_doppelganger(
+    "geom_likert and geom_likert_text",
+    p
+  )
+
+  d <- Titanic |> as.data.frame()
+  p <- ggplot(d) +
+    aes(y = Class, fill = Sex, weight = Freq) +
+    geom_pyramid() +
+    geom_pyramid_text()
+  vdiffr::expect_doppelganger(
+    "geom_pyramid and geom_pyramid_text",
     p
   )
 })